util/sendSummary -> announce.heartbeat
