+++ community-contributions/joystreamtelegrambot/config.ts
@@ -10,8 +10,8 @@ export const token: string = "";
 // telegram chat ID
 export const chatid: string = "";
 
-// time period between summaries in milliseconds
-export const summaryPeriod = 10000 * 60 * 24;
+// time between heartbeat announcement in milliseconds
+export const heartbeat = 60000 * 60 * 24;
 
 /// account to query staking
 export const accountId = "";
